refactor(p2p): extract remote topology operations
